requirements/base.txt
pip==8.0.0
setuptools==19.4

Django==1.9.1
psycopg2==2.6.1
Pillow==3.1.0
django-crispy-forms==1.6
django-model-utils==2.4
django-reversion==1.10.0
djangorestframework==3.3.2
django-appconf==1.0.1
django_compressor==2.0
libsass==0.10.0
django-libsass==0.6
git+https://github.com/jbsag/django-activity-stream.git@395e892e27784cf8cdd01f9a6c8966fdd548b587#egg=django-activity-stream
markdown2==2.3.0
Pygments==2.1
bleach==1.4.2
html5lib==0.9999999
beautifulsoup4==4.4.1
colour==0.1.2
requests==2.9.1
dukpy-lukegb==0.2.2
git+https://github.com/lukegb/Jawa.git@c0dde132389f167c382d29211d42055b2ccfeecc#egg=Jawa
parsimonious==0.6.2
